int 改 long 数据质量超int范围

+++ biz-module-flink-pro/src/main/java/com/yqwl/datamiddle/realtime/bean/DwdSptb02.java
@@ -601,7 +601,7 @@ public class DwdSptb02 implements Serializable {
      * 返单大序号  number 转 int
      */
 
-    private Integer NFDDXH;
+    private Long NFDDXH;
 
     /**
      * 发车站台
