+++ biz-module-flink-pro/src/main/java/com/yqwl/datamiddle/realtime/bean/Sptb02.java
@@ -593,7 +593,7 @@ public class Sptb02 implements Serializable {
      * 返单大序号  number 转 int
      */
 
-    private Integer NFDDXH;
+    private Long NFDDXH;
 
     /**
      * 发车站台
